chore: use standard package name conventions
