+++ build.xml
@@ -22,7 +22,7 @@
       basedir="${classes.dir}"
     >
       <manifest>
-        <attribute name="Main-Class" value="Main" />
+        <attribute name="Main-Class" value="com.jumbosoft.paint.Main" />
       </manifest>
     </jar>
   </target>
